refactor: visitors
